% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_no_rescue.R
\name{cand_no_rescue}
\alias{cand_no_rescue}
\title{Recover candidate number input}
\usage{
cand_no_rescue(filepath)
}
\arguments{
\item{filepath}{a path to the html TAP file}
}
\value{
candidate number input (character vector)
}
\description{
Recover candidate number input
}
\examples{
 \dontrun{
 cand_no_rescue("path/to/TAP.html")
}
}
