% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{fr}
\alias{fr}
\title{Format-round helper}
\usage{
fr(x, digits = 3)
}
\arguments{
\item{x}{a number to format for reporting}

\item{digits}{number of digits}
}
\value{
a rounded character object
}
\description{
Format-round helper
}
\examples{
fr(0.0123423, 2)
}
