% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mousetrap_checkr.R
\name{mousetrap_checkr}
\alias{mousetrap_checkr}
\title{Run a shiny app to check TAP results}
\usage{
mousetrap_checkr(...)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
runs a shiny app
}
\description{
Run a shiny app to check TAP results
}
\examples{
 \dontrun{
 mousetrap_checkr()
}
}
